+++ sources/init_matrix.c
@@ -6,7 +6,7 @@
 /*   By: dmalesev <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 14:05:22 by dmalesev          #+#    #+#             */
-/*   Updated: 2022/10/14 12:54:21 by dmalesev         ###   ########.fr       */
+/*   Updated: 2022/10/15 10:13:56 by dmalesev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,43 +26,43 @@ t_mat	init_pmatrix(t_proj *proj)
 	return (matrix);
 }
 
-t_mat	init_rmatrix_x(float angle_x)
+t_mat	init_rmatrix_x(double angle_x)
 {
 	t_mat	matrix;
 
 	ft_bzero(&matrix, sizeof(t_mat));
 	matrix.m[0][0] = 1;
-	matrix.m[1][1] = (float)cos(angle_x / 180.0f * PI);
-	matrix.m[1][2] = (float)sin(angle_x / 180.0f * PI);
-	matrix.m[2][1] = (float)-sin(angle_x / 180.0f * PI);
-	matrix.m[2][2] = (float)cos(angle_x / 180.0f * PI);
+	matrix.m[1][1] = cos(angle_x / 180.0f * PI);
+	matrix.m[1][2] = sin(angle_x / 180.0f * PI);
+	matrix.m[2][1] = -sin(angle_x / 180.0f * PI);
+	matrix.m[2][2] = cos(angle_x / 180.0f * PI);
 	matrix.m[3][3] = 1;
 	return (matrix);
 }
 
-t_mat	init_rmatrix_y(float angle_y)
+t_mat	init_rmatrix_y(double angle_y)
 {
 	t_mat	matrix;
 
 	ft_bzero(&matrix, sizeof(t_mat));
-	matrix.m[0][0] = (float)cos(angle_y / 180.0f * PI);
-	matrix.m[0][2] = (float)-sin(angle_y / 180.0f * PI);
+	matrix.m[0][0] = cos(angle_y / 180.0f * PI);
+	matrix.m[0][2] = -sin(angle_y / 180.0f * PI);
 	matrix.m[1][1] = 1;
-	matrix.m[2][0] = (float)sin(angle_y / 180.0f * PI);
-	matrix.m[2][2] = (float)cos(angle_y / 180.0f * PI);
+	matrix.m[2][0] = sin(angle_y / 180.0f * PI);
+	matrix.m[2][2] = cos(angle_y / 180.0f * PI);
 	matrix.m[3][3] = 1;
 	return (matrix);
 }
 
-t_mat	init_rmatrix_z(float angle_z)
+t_mat	init_rmatrix_z(double angle_z)
 {
 	t_mat	matrix;
 
 	ft_bzero(&matrix, sizeof(t_mat));
-	matrix.m[0][0] = (float)cos(angle_z / 180.0f * PI);
-	matrix.m[0][1] = (float)sin(angle_z / 180.0f * PI);
-	matrix.m[1][0] = (float)-sin(angle_z / 180.0f * PI);
-	matrix.m[1][1] = (float)cos(angle_z / 180.0f * PI);
+	matrix.m[0][0] = cos(angle_z / 180.0f * PI);
+	matrix.m[0][1] = sin(angle_z / 180.0f * PI);
+	matrix.m[1][0] = -sin(angle_z / 180.0f * PI);
+	matrix.m[1][1] = cos(angle_z / 180.0f * PI);
 	matrix.m[2][2] = 1;
 	matrix.m[3][3] = 1;
 	return (matrix);
